% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age_groups.R
\name{add_age_groups}
\alias{add_age_groups}
\title{Title Add age group column to data.frame}
\usage{
add_age_groups(df, age_column, age_groups, intermediates, unknowns, factor)
}
\arguments{
\item{df}{data.frame with an age column}

\item{age_column}{column from which to calculate ages}

\item{age_groups}{Input vector of age ranges with upper range marked as x+}

\item{intermediates}{values which could go up or down (under development)}

\item{unknowns}{vector of Age values to be considered unknowns eg -1, NA, "not stated"}

\item{factor}{}
}
\value{
Returns input data.frame with added age group column
}
\description{
Title Add age group column to data.frame
}
\examples{

unknowns <- c(-1, NA)
age_groups <- c('0-4', "5-10", "11-16", "17-24", "25-30", "31-39", "40-59", "60-79", "80+")
age_column <- "Age"
factor <- F
intermediates <- "down"
add_age_groups(df, age_column, age_groups, intermediates, unknowns, factor)

}
