% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age_groups.R
\name{age_groups}
\alias{age_groups}
\title{Title Add age group column to data.frame}
\usage{
age_groups(
  age,
  groups = c("0-4", "5-10", "11-16", "17-24", "25-30", "31-39", "40-59", "60-79",
    "80+"),
  unknowns = c("-1", NA),
  factor = FALSE
)
}
\arguments{
\item{age}{age vector}

\item{groups}{Input vector of age ranges in format "N1-N2" where N1 is the lower age and N2 the upper age. Mark range where max is infinite as "N1+"}

\item{unknowns}{vector of Age values to be considered unknowns eg -1, NA, "not stated"}

\item{factor}{}
}
\value{
Returns age groups
}
\description{
Title Add age group column to data.frame
}
\examples{
# df <- data.frame(ID = c(1:4), Age = c(8,-1,38,26))
# groups <- c('0-4', "5-10", "11-16", "17-24", "25-30", "31-39", "40-59", "60-79", "80+")
# df$group <- age_groups(df$Age, groups, unknowns, factor)

}
