% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive}
\alias{archive}
\title{Archives files from a folder based on modification date.
All files within folder are processed based on modification date.}
\usage{
archive(from, to, date, dir, string, exclude, keep)
}
\arguments{
\item{from}{Path to folder containing files to archive.}

\item{to}{Folder name to archive files to. Default is Archive}

\item{date}{Minimum modification date to exclude files from archiving. Default is Sys.Date().}

\item{dir}{Optional logical argument; also archive folders or not. Default is FALSE.}

\item{string}{Optional string pattern in files to archive.}

\item{exclude}{Optional string pattern of files to exclude from archiving}

\item{keep}{Optional retain archived files in original location and copy to archive folder. Default is FALSE.}
}
\value{
Moves older files based on date to an archive subfolder
}
\description{
Archives files from a folder based on modification date.
All files within folder are processed based on modification date.
}
\examples{

# Move all files excluding folders to an Archive subfolder
archive("C:..../analysis")

# Move all files excluding folders to a different folder
archive(from = "C:..../analysis", to = "C:.../backup/outputs")

# Copy all files and folders with the "output_version_x" in the file name which have not been modified in the past week to a different folder
archive(from = "C:..../analysis", to = "C:.../backup/outputs", date = Sys.Date() -7,
        string = "output_version_x", keep = T, dir = T)


}
\author{
Diane Hatziioanou
}
\keyword{archive}
