% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_position.R
\name{axis_number}
\alias{axis_number}
\title{Title Round large numbers up or increase by certain value at defined position}
\usage{
axis_number(x, fromleft = 2, value = 1)
}
\arguments{
\item{x}{numbers}

\item{fromleft}{position within number from left to round up}
}
\value{
x rounded up at defined position
}
\description{
Title Round large numbers up or increase by certain value at defined position
}
\examples{

x <- c(290, 5218, 462964)
y_max <- axis_number(max(x))

}
