% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_manage.R
\name{col_namecompare}
\alias{col_namecompare}
\title{Title Column names list duplicates in 1 or more objects. Disregards case.}
\usage{
col_namecompare(x)
}
\arguments{
\item{x}{list of objects}
}
\value{
Returns list with same = duplicates in same object, other = duplicates between objects
}
\description{
Title Column names list duplicates in 1 or more objects. Disregards case.
}
\examples{
# colnamecheck(df)
# colnamecheck(list(df1, df2, df3))

}
