% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_names.R
\name{col_nameregex}
\alias{col_nameregex}
\title{Title Column name search for string patterns}
\usage{
col_nameregex(dt, strings, out = "names")
}
\arguments{
\item{dt}{object such as dataframe, datatable, list}

\item{strings}{one or more strings to search for}

\item{out}{return column names or positions}
}
\value{
returns column names or positions
}
\description{
Title Column name search for string patterns
}
\examples{
# col_nameregex(dt,"dob", out = "names")

}
