% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnamecheck.R
\name{colnamecheck}
\alias{colnamecheck}
\title{Title List duplicate column names in 1 or more objects. Disregards case.}
\usage{
colnamecheck(x)
}
\arguments{
\item{x}{list of objects}
}
\value{
Returns list with same = duplicates in same object, other = duplicates between objects
}
\description{
Title List duplicate column names in 1 or more objects. Disregards case.
}
\examples{
colnamecheck(df)
colnamecheck(list(df1, df2, df3))

}
