% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_manage.R
\name{combine_vectors}
\alias{combine_vectors}
\title{Title Combine vectors or table columns while keeping all values.}
\usage{
combine_vectors(l, r, case = NULL)
}
\arguments{
\item{l}{left hand side column}

\item{r}{right hand side column}

\item{case}{Default of NULL does not formatting. Options upper, lower or title format data before combining.}
}
\value{
returns vector of data combined by position
}
\description{
Title Combine vectors or table columns while keeping all values.
}
\examples{
# columns_combine(df$unique_id, df$forename)

}
