% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\name{episodes}
\alias{episodes}
\title{Add episode information based on the duration of episodes}
\usage{
episodes(dt, unique_column, time_column, duration)
}
\arguments{
\item{unique_column}{Unique identifier column to summarise data by (string)}

\item{time_column}{Date column to summarise data by}

\item{duration}{Numeric duration of episodes in days}

\item{data}{input data}
}
\value{
Returns a data.table with summary
sample_time_diff = time difference between samples
New_episode = 1 for new episodes, 0 for repeat samples from the same episode.
episodeN = episode number for unique identifier
episode_time_diff = time difference from previous sample within episode
}
\description{
Add episode information based on the duration of episodes
}
\examples{

}
