% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\name{extract_fasta}
\alias{extract_fasta}
\title{Title Extract fasta from file in a loop}
\usage{
extract_fasta(
  filepath,
  batch = 1000,
  skip = 0,
  returnids = TRUE,
  fastaids = vector(),
  fastafileout = "fastaout.txt",
  idfileout = "fastaindex.txt"
)
}
\arguments{
\item{filepath}{fasta file to search}

\item{batch}{number of rows to read at a time}

\item{skip}{number of rows to skip from fasta file}

\item{returnids}{logical, TRUE to return identifiers}

\item{fastaids}{fasta identifiers to retrieve}

\item{fastafileout}{file name to save fasta matching ids to}

\item{idfileout}{file name to save fasta index to}
}
\value{
saves identifier matches fasta to file and returns a list of number of ids searched, last row searched, ids foundand their row numbers, fasta ids searched for,the file names for the fasta id matches and the fasta index.
}
\description{
Title Extract fasta from file in a loop
}
\examples{
extract_fasta(filepath="audit.fasta")

}
