% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_vc.R
\name{file_stat_diff}
\alias{file_stat_diff}
\title{Title}
\usage{
file_stat_diff(
  logs = file_stat_log,
  columns,
  maxup = 1.25,
  maxdown = 0,
  logpath = NULL,
  logname = NULL,
  Date = Sys.Date()
)
}
\arguments{
\item{logs}{log created using file_stat_log}

\item{columns}{columns within log with quantified value combinations}

\item{maxup}{Maximum permissible increase from previous log. Warning recorded if exceeded.}

\item{maxdown}{Maximum permissible decrease from previous log. Warning recorded if exceeded.}
}
\value{

}
\description{
Title
}
