% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_vc.R
\name{file_stat_log}
\alias{file_stat_log}
\title{Title Create and maintain a record of the number of rows in an input with each combination of values from columns of interest}
\usage{
file_stat_log(
  logpath = NULL,
  logname = NULL,
  input,
  inputname,
  columns,
  Date = Sys.Date()
)
}
\arguments{
\item{logpath}{file path where log is saved.}

\item{logname}{string containing descriptive name of log. Processing date will be added to this after creation.}

\item{input}{Cumulative dataset to be processed.}

\item{columns}{columns within input with value combinations to quantify.}
}
\value{

}
\description{
Title Create and maintain a record of the number of rows in an input with each combination of values from columns of interest
}
