% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_of_rows_by_geography_time.R
\name{frequency_of_rows_by_geography_time}
\alias{frequency_of_rows_by_geography_time}
\title{Title Calculate frequency of line list row counts by geography and date variables}
\usage{
frequency_of_rows_by_geography_time(data, geography_column, time_column)
}
\arguments{
\item{data}{line list with a geography and date column}

\item{geography_column}{Variable with geography}

\item{time_column}{Variable with dates}
}
\value{
returns data with additional EMA, SMA and change variables
}
\description{
Title Calculate frequency of line list row counts by geography and date variables
}
