% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuz_search}
\alias{fuz_search}
\title{Title}
\usage{
fuz_search(dbaob, sn_fun)
}
\arguments{
\item{dbaob}{database to search against}

\item{sn_fun}{full name to search with all combinations of all names and DOB}
}
\value{

}
\description{
Title
}
\examples{
  # Retrieve db name matches
  matches  <- fuz_search(dbaob, sn_fun)
}
