% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Fuzzy match a custom line list to a dabatase by Suname, first name and DOB}
\usage{
fuzzy_match(lla, dba, stringency, pc_format, case_column, parallel)
}
\arguments{
\item{lla}{Input Line-list list containing the items ob, sn, fn, dob, pc, geo}

\item{dba}{Input database list containing the items ob, sn, fn, dob, pc, geo}

\item{stringency}{Optional; stringency required stringency of matching; one of high, medium or low. Default is high.}

\item{pc_format}{postcode format where l = letters and n = numbers. Default is the UK format "lnl"}

\item{case_column}{Column within the custom line list with one unique case number per row}

\item{parallel}{Optional; process in parallel. Default is TRUE}

\item{ob}{database object to search. All columns of this database will be appended to the input line list}

\item{sn}{surname column within the database object as a string}

\item{fn}{first name column within the database object as a string}

\item{dob}{date of birth column within the database object as a string}

\item{pc}{postcode column within the database object as a string}

\item{geo}{Geography column containing the location in string format from lla geo where most individuals can be found.}
}
\value{
Returns the original input object as a data.frame with added columns of the input databse where there is a match base on full name and DOB
}
\description{
Matches custom line list persons by full name and DOB to a database, assigns scores and creates warnings based on the match quality and returns the original line list with all the columns of the database and database information where there is a match to the custom line list
}
\author{
Diane Hatziioanou
}
