% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlatestfile.R
\name{getlatestfile}
\alias{getlatestfile}
\title{Retrieve most recent file from a file path}
\usage{
getlatestfile(folder_path, file_string, exclusions, return_type, maxTries)
}
\arguments{
\item{folder_path}{Folder path containing the file to retrieve. Folders within this path will not be searched.}

\item{file_string}{Character string within the file names to retrieve (optional). If not defined all file names within the folder are retrieved.}

\item{exclusions}{Character string within the file names for file exclusion (optional). If not defined nothing is excluded from the retrieved file name list.}

\item{return_type}{Use "all" to retrieve a data.frame with the file name, file path and modiciation date, "name" to retrieve the file name or "path" to retrieve the full file path. Default is "path". (optional)}

\item{maxTries}{Maximum number of times to attempt file information retrieval. Default is 1. (optional)}
}
\value{
Returns the last ctime file path, file name or a data.frame with the file name, path and modification details as defined by return_type where size is the file size in bytes, mtime is the last modification time, ctime is last status change time, atime is last access time.
}
\description{
Retrieves the latest file name and modification details  for a specified path with the option to only look include files with a defined name string and exclude files with other defined name strings. The function can be set to attempt file retrieval a defined number of times to overcome server connectivity problems.
}
\examples{
Return a path
file_path <- getlatestfile(file.path("...","subfolder","subfolder"))

Import the latest csv file
df <- fread(getlatestfile(folder_path = file.path("...","subfolder","subfolder"), file_string = "csv",exclusions = "unwanted", return_type = "path", maxTries = 5))

Record the modification details
Data_as_of <- getlatestfile(folder_path = file.path("...","subfolder","subfolder"), file_string = "csv",exclusions = "unwanted", return_type = "all", maxTries = 5)$ctime

}
\author{
Diane Hatziioanou
}
\keyword{file}
\keyword{name,}
\keyword{path}
