% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDinfilesearch.R
\name{IDinfilesearch}
\alias{IDinfilesearch}
\title{Title Search a list of files for values in a consistent column}
\usage{
IDinfilesearch(IDs, IDcol, files, complete = FALSE)
}
\arguments{
\item{IDs}{Id values to look for.}

\item{IDcol}{Column containing IDs in files.}

\item{files}{Files to look in}

\item{complete}{Optional; retain files without matches in list or not. Default is FALSE}
}
\value{
Returns a list with file names and rows of data with listed IDs in the IDCol column
}
\description{
Title Search a list of files for values in a consistent column
}
\examples{

# simplewrite(data.frame(a = c(1,2,3), b = c("red", "amber", "green")), "testa.rds")
# simplewrite(data.frame(a = c(1,2,3, 4), b = c("red", "amber", "green", "blue")), "testb.rds")
# files <- list.files(pattern = ".rds")
# dt3 <- IDinfilesearch(IDs = c(3,4), IDcol = "a", files = files)
# dt4 <- IDinfilesearch(IDs = 4, IDcol = "a", files = files, complete = TRUE)

}
