% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplededup.R
\name{name_combos}
\alias{name_combos}
\title{Title Format names from a data.table and make return list of all name combinations per row.}
\usage{
name_combos(x, namecols)
}
\arguments{
\item{x}{data.table of names}

\item{namecols}{names of column within x with names to process}
}
\value{
list of all name combinations
}
\description{
Title Format names from a data.table and make return list of all name combinations per row.
}
\examples{
 dt <- data.table(first_name = c("Jane", "Pete"), surname = c("Brown", "McKenzie"))
 dt$namelist <- name_combos(x = dt, namecols =  c("first_name","surname"))

}
