% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONS_age_groups.R
\name{ONS_age_groups}
\alias{ONS_age_groups}
\title{Title Calculate ONS populations by age group for any age grouping using the ONS populations file}
\usage{
ONS_age_groups(populations, age_groups)
}
\arguments{
\item{populations}{object containing ONS populations}

\item{age_groups}{data.frame containing age group categories to calculate populations for}
}
\value{

}
\description{
Title Calculate ONS populations by age group for any age grouping using the ONS populations file
}
\examples{


 Import population data
 ONS_pop <- readxl::read_xlsx(file.path(...,  "SAPE21DT1a-mid-2018-on-2019-LA-lsoa-syoa-estimates-formatted.xlsx"), sheet = "Mid-2018 Persons",  col_names  =  T, skip = 4)

 Create age grouping
 age_groups_3cat <- data.frame(
   Population_under18 = c(0,17),
   Population_workingage = c(18,65),
   Population_over65 = c(66,"90+"),
   stringsAsFactors = F)

 age_groups_9cat <- data.frame(
   `Population_0-4yrs` = c(0,4),
   `Population_5-10yrs` = c(5,10),
   `Population_11-17yrs` = c(11,17),
   `Population_18-29yrs` = c(18,29),
   `Population_30-39yrs` = c(30,39),
   `Population_40-49yrs` = c(40,49),
   `Population_50-59yrs` = c(50,59),
   `Population_60-69yrs` = c(60,69),
   `Population_70+yrs` = c(70,"90+"),
   stringsAsFactors = F, check.names = T, fix.empty.names = T)

 ONS_populations_3categories <- ONS_age_groups(ONS_pop, age_groups_3cat)
 ONS_populations_9categories <- ONS_age_groups(ONS_pop, age_groups_9cat)
}
