% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_categories.R
\name{QC_categories}
\alias{QC_categories}
\title{Title}
\usage{
QC_categories(data, cat)
}
\arguments{
\item{data}{vector to perform check on}

\item{cat}{categorical values to look for}
}
\value{

}
\description{
Title
}
\examples{

# check1 <- QC_categories(df$coolumn1, cat = c("low", "medium", "high"))

}
