% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_replicates.R
\name{remove_replicates}
\alias{remove_replicates}
\title{Columns aggregated with non unique comma separated}
\usage{
remove_replicates(df, columns)
}
\arguments{
\item{df}{data}

\item{columns}{column or vector of columns to group by}
}
\value{
Returns df aggregated by the grouping columns with non unique values comma separated
}
\description{
Aggregate a dataset such as a line list by a number of grouping variables and comma separate non-unique values in the rest
}
\examples{
df <- remove_replicates(df, c("Name","DOB"))

df <- remove_replicates(df, c("Name","DOB", "episode"))

}
\author{
Diane Hatziioanou
}
\keyword{aggregate,}
\keyword{replicates}
