% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_manage.R
\name{replicates}
\alias{replicates}
\title{Subset data.frame or data.table for all rows duplicated in one or more columns}
\usage{
replicates(df, columns)
}
\arguments{
\item{df}{data.frame or data.table}

\item{columns}{column or vector of columns to look for duplicate combinations in}
}
\value{
Returns duplicated rows within df based on the specified columns as data.frame
}
\description{
Subset data.frame or data.table for all rows duplicated in one or more columns
}
\examples{
# df_duplicates <- replicates(df, c("Name","DOB"))


}
\author{
Diane Hatziioanou
}
\keyword{duplicates,}
\keyword{replicates}
