% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_manage.R
\name{replicates_rmv}
\alias{replicates_rmv}
\title{Subset data.frame or data.table to remove all duplicated rows by one or more columns}
\usage{
replicates_rmv(df, columns)
}
\arguments{
\item{df}{data.frame or data.table}

\item{columns}{column or vector of columns to look for duplicate combinations in}
}
\value{
Returns data.frame without duplicated rows based on the specified columns
}
\description{
Subset data.frame or data.table to remove all duplicated rows by one or more columns
}
\examples{

# df_exclude_all_duplicates <- replicates_rmv(df, c("ID1","ID2"))


}
\author{
Diane Hatziioanou
}
\keyword{duplicates,}
\keyword{replicates}
