% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Title}
\usage{
retrieve_data(i, case_column, llaob, dbaob)
}
\arguments{
\item{i}{Cycle number}

\item{case_column}{column with unique identifier}

\item{llaob}{line list database}

\item{dbaob}{search datavase}
}
\value{

}
\description{
Title
}
\examples{

 # par_list <- retrieve_data(i, case_column, llaob, dbaob)

}
