% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{seq.yearweek}
\alias{seq.yearweek}
\title{Title Create a year-week sequence}
\usage{
\method{seq}{yearweek}(from, to, by, rangef = "yearweek")
}
\arguments{
\item{from}{Start date in yearweek (YYYYWW) or date format}

\item{to}{End data in yearweek (YYYYWW) or date format}

\item{by}{Increments for sequence in weeks}

\item{rangef}{Format of time-range (from and to); one of yearweek or date.Default is yearweek.}
}
\value{
Returns a sequence of yearweek values for the range
}
\description{
Title Create a year-week sequence
}
\examples{
# seq.yearweek(from = 202102, to = 202105, by = 1, rangef = "yearweek")
# seq.yearweek(from = "20210205", to = "20210510", by = 2, rangef = "date")


}
