% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpledates.R
\name{simple_dates}
\alias{simple_dates}
\title{Format dates}
\usage{
simple_dates(x)
}
\arguments{
\item{x}{dates to format. Takes a single date, vectors or a column of data.}
}
\value{

}
\description{
Format dates either from a string format or an excel exported numeric format
}
\examples{
# Format a single string into a date
date <- fix_dates("2020739")

# Format a data.frame column into a consitent date format
df$column <- fix_dates(df$column)

# Format all column from within a data.table where the columns contain the string date into date format
for (col in (names(dt)[grepl("date", names(dt))])){dt[,(col) := simpledates(dt[,get(col)])]}

}
