% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_manage.R
\name{simple_format_calatogue}
\alias{simple_format_calatogue}
\title{Title Create table of data.table column characteristics}
\usage{
simple_format_calatogue(dt)
}
\arguments{
\item{dt}{data.table. If a dataframe is provided it will be converted to a datatable.}
}
\value{
returns a data.table with column length metrics min, max, mean, median, na metrics, class, skewness and kurtosis per column.
}
\description{
Title Create table of data.table column characteristics
}
\examples{
# formats_dt <- simple_format_calatogue(dt)


}
