% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_vc.R
\name{simple_vc}
\alias{simple_vc}
\title{Title Compare two versions of data with the same length and mark new data in a version control column}
\usage{
simple_vc(
  oldcol = dt$oldcol,
  newcol = dt$newcol,
  vccol = dt$vccol,
  olddate = "Original",
  newdate = Sys.Date()
)
}
\arguments{
\item{oldcol}{vector of previous data}

\item{newcol}{vector of new data}

\item{vccol}{existing or new vector to record changes in}

\item{olddate}{date of previous data for version control}

\item{newdate}{date of new data for version control}
}
\value{
returns vccol newdate with new changes
}
\description{
Title Compare two versions of data with the same length and mark new data in a version control column
}
\examples{
 dt =data.table(oldcol = c("A", "B", "C", NA, "D", "", "Test"),
               newcol = c("", "Test", NA, NA, "C", NA, "newdata"))
 dt$vc = simple_vc(oldcol = dt$oldcol, newcol = dt$newcol, vccol = dt$vccol)

 dt$vccol = c("Original;A", "Original;B", " Original;C",  "Original;C then NA", " Original;D", NA, "no previous")
 dt$vcupdate = simple_vc(oldcol = dt$oldcol, newcol = dt$newcol, vccol = dt$vccol)

}
