% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleappendenv.R
\name{simpleappendenv}
\alias{simpleappendenv}
\title{Title Append object data in environment throughout loop}
\usage{
simpleappendenv(x, name = "DF")
}
\arguments{
\item{x}{data to append during a loop}

\item{name}{name to call database of appended x}
}
\value{
object in .GlobalEnv containing data from x with new data from itterations of the code appended at the end
}
\description{
Title Append object data in environment throughout loop
}
\examples{
for (v in c(1,2,3)){
  df <- data.frame(a = stats::runif(3), b = c("a", "b", "c"))
  df$version <- v
  simpleappend(df, name = "DF")
}

}
