% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpledates.R
\name{simpledates}
\alias{simpledates}
\title{Format dates}
\usage{
simpledates(x, char = NA, silent = TRUE)
}
\arguments{
\item{x}{dates to format. Takes a single date, vectors or a column of data.}

\item{char}{value to return for character strings. Use historic date values to manage categories where date not known but label required for downstream processing.}
}
\value{
Returns input formatted as Date
}
\description{
Format dates either from a string format or an excel exported numeric format
}
\examples{
# Format a single string into a date
# date <- simpledates("2020739")

# Format a data.frame column into a consistent date format
# df$column <- simpledates(df$column)

# Format all column from within a data.table where the columns contain the string date into date format
# for (col in (names(dt)[grepl("date", names(dt))])){
#   dt[,(col) := simpledates(dt[,get(col)])]
#  }
}
