% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplededup.R
\name{simplededup}
\alias{simplededup}
\title{Title Manage duplicates in line list}
\usage{
simplededup(
  x,
  namecols = c("first_name", "surname"),
  dobcol = "date_of_birth",
  pccol = "postcode",
  idcols = c("nhs_number"),
  pc_format = "lnl"
)
}
\arguments{
\item{x}{Data}

\item{namecols}{name columns}

\item{dobcol}{dob column}

\item{pccol}{postcode column}

\item{idcols}{id columns}

\item{pc_format}{postcode format}
}
\value{

}
\description{
Title Manage duplicates in line list
}
\examples{
# simplededup()

}
