% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleencode.R
\name{simpleencode}
\alias{simpleencode}
\title{Title Change character encoding to UTF-8 with optional case management}
\usage{
simpleencode(x, case = "none")
}
\arguments{
\item{x}{Character object or vector to encode.}

\item{case}{Optional; Default is none (no formatting). Other options are upper, lower or title.}
}
\value{
returns the input encoded to UTF-8
}
\description{
Title Change character encoding to UTF-8 with optional case management
}
\examples{
x <- "troublesome character string"
x <- simpleencode(x)

}
