% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplefilecheck.R
\name{simplefilecheck}
\alias{simplefilecheck}
\title{Title Check if two files are present and of the same size and same creation and modification time}
\usage{
simplefilecheck(x, y)
}
\arguments{
\item{x}{file 1 to check}

\item{y}{file 2 to check}
}
\value{
TRUE if both files exist, FALSE if the files are not identical
}
\description{
Title Check if two files are present and of the same size and same creation and modification time
}
\examples{

simplewrite(list("test"), "file1.csv")
simplewrite(list("test"), "file2.csv")
simplefilecheck("file1.csv", "file2.csv")

}
