% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleID.R
\name{simpleID}
\alias{simpleID}
\title{Title Create a unique identified based on unique values in columns and a further unique identifier
Assign unique IDs to data based either on PII columns or a combination of PII columns and a further unique identifier where available}
\usage{
simpleID(x, col, Number, name)
}
\arguments{
\item{x}{Data to assign unique IDs to}

\item{col}{PII columns within data to use for assigning unique IDs}

\item{Number}{Potentially incomplete unique identifier which can be used in combination with PII to assign unique IDs}

\item{name}{Name to give unique ID column in output}
}
\value{
full dataset with additional ID column returned
}
\description{
Title Create a unique identified based on unique values in columns and a further unique identifier
Assign unique IDs to data based either on PII columns or a combination of PII columns and a further unique identifier where available
}
\examples{

x <- data.frame(DOB = c("01/01/2001", "02/01/2001", "01/01/2001", "03/01/2001"), name = c("A", "B", "C", "D"), surname = c("A", "B", "A", "D"), ZIP = c("A", "B", "A", "D"), NHS_Number = c(2001, NA, 2001, NA))
Number = "NHS_Number"
name <- "GroupID"
col = c("DOB", "name", "surname", "ZIP")
simpleID(x = x, col =  col, name = name)
simpleID(x = x, col =  col, Number = Number, name = name)
}
