% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{simpleimport}
\alias{simpleimport}
\title{Import various formats of files into R with one command. If file is not found a brower will open to select the file}
\usage{
simpleimport(file, sheet, skip, password = NULL, ...)
}
\arguments{
\item{file}{Optional; file to import. If not in your working directory the path needs to be included. If not specified a browser will open to select a file.}

\item{sheet}{Optional; sheet name or number within workbooks to import. By default the first tab is imported.}

\item{skip}{Optional; Number of rows from top of the spreadsheet to skip in the import.}

\item{password}{Password for password-protected workbook.}

\item{...}{Optional; other import arguments from the fread for csv, readxl for xls or xlsx or haven for dta files.}
}
\value{
files imported into R
}
\description{
Quick import of file types Rdata, RDS, rda, csv, txt, xlsx, xls or dta into your R environment.
}
\examples{

# simplewrite(data.frame(a = c(1,2,3), b = c("red", "amber", "green")), "test")
# df <- simpleimport("test.csv")

# simplewrite(data.frame(a = c(1,2,3), b = c("red", "amber", "green")), "test.xlsx")
# df <- simpleimport("test.xlsx")


# simplewrite(data.frame(a = c(1,2,3), b = c("red", "amber", "green")), "test.rds")
# df <- simpleimport("test.rds")

}
\author{
Diane Hatziioanou
}
\keyword{import}
