% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleIsoweek.R
\name{simpleIsoweek}
\alias{simpleIsoweek}
\title{Title Get Iso year-week from a date object}
\usage{
simpleIsoweek(x, integer)
}
\arguments{
\item{x}{Date value or vector to convert to Iso year-week}

\item{integer}{Return an integer instead of a cahracter vector. Default is FALSE}
}
\value{
Returns Iso year-weeks as strings in the format YYYYWW
}
\description{
Title Get Iso year-week from a date object
}
\examples{
df$Isoweek <- simpleIsoweek(df$sample_date)

}
