% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplelog.R
\name{simplelog_end}
\alias{simplelog_end}
\title{Title Write data and end connection to history and log files}
\usage{
simplelog_end(path, history_file, output_file)
}
\arguments{
\item{path}{Optional path to history and output files. Default is working directory.}

\item{history_file}{Optional file name for history. Default is file.path(path, paste0(date,"_Rhistory.txt"))}

\item{output_file}{Optional file name for output. Default is file.path(path, paste0(date,"_Rlog.txt"))}
}
\value{
Closes connection to history and output files
}
\description{
Title Write data and end connection to history and log files
}
\examples{
# simplelog_end(path, history_file, output_file)

}
