% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplelog.R
\name{simplelog_start}
\alias{simplelog_start}
\title{Title Start connection to history and log files}
\usage{
simplelog_start(path, history_file, output_file)
}
\arguments{
\item{path}{Path to history and output files}

\item{history_file}{file name for history}

\item{output_file}{file name for output}
}
\value{
Opens connection to history and output files
}
\description{
Title Start connection to history and log files
}
\examples{
simplelog_start()

}
