% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplenumber.R
\name{simplenumber}
\alias{simplenumber}
\title{Title
Remove letters and special characters from a string to return a numeric value}
\usage{
simplenumber(x, fillempty, fillvalue)
}
\arguments{
\item{x}{string or vector with numeric values for cleaning}
}
\value{
number or vector of numbers
}
\description{
Title
Remove letters and special characters from a string to return a numeric value
}
\examples{
# x <- c("dfju", "LASK3", "dfpisdu092", "9376", 3098)
# x <- simplenumber(x, fillempty = T, fillvalue = -1)

}
