% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleopenfolder.R
\name{simpleopen}
\alias{simpleopen}
\title{Title Open File Explorer at a specified path or open files from R.}
\usage{
simpleopen(path)
}
\arguments{
\item{path}{Path to folder or file to open}
}
\value{
Opens a File Explorer window at the given path or file
}
\description{
Title Open File Explorer at a specified path or open files from R.
}
\examples{
# dir.create("test")
# simpleopen("test")

}
