% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleopenfolder.R
\name{simpleopenfolder}
\alias{simpleopenfolder}
\title{Title Open File Explorer at a specified path. Only tested for Windows}
\usage{
simpleopenfolder(path)
}
\arguments{
\item{path}{Path to open in a File Explorer}
}
\value{
Opens a File Explorer window at the given path
}
\description{
Title Open File Explorer at a specified path. Only tested for Windows
}
\examples{
simpleopenfolder(file.path("C:","Program Files","R"))

}
