% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_manage.R
\name{simpleunzip}
\alias{simpleunzip}
\title{Title Unzip a file using 7z}
\usage{
simpleunzip(zfile, z7path)
}
\arguments{
\item{zfile}{zip file to unzip}

\item{z7path}{path to zip file}
}
\value{
unzipped zip file contents
}
\description{
Title Unzip a file using 7z
}
\examples{
#simpleunzip(zfile, z7path)

}
