% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplewords.R
\name{simplewords}
\alias{simplewords}
\title{Remove special characters and numbers and format to UTF-8 encoding with standardised capitalisation}
\usage{
simplewords(x, case)
}
\arguments{
\item{x}{string or character vector}

\item{case}{Case of words to return; one of upper, lower or title for the first letter uppercase and the rest lower case}
}
\value{
Returns the input in UTF-8 format with special characters replaced with spaces and numbers removed.
}
\description{
Takes a vector with character elements, removes special characters and numbers and formats them to UTF-8 encoding to make these consistent and compatible with most R analysis procedures. Returns strings with the first letter capitalised and the rest in lower case
}
\examples{
strings <- tidywords(strings)

clean_name <- tidywords("Latin-multi^name with strange/characters")

df$Name <- tidywords(df$Name, case = "lower")


}
\author{
Diane Hatziioanou
}
\keyword{characters}
\keyword{encoding,}
\keyword{special}
