% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{simplewrite}
\alias{simplewrite}
\title{Write data to file in spreadsheet format}
\usage{
simplewrite(
  data,
  file,
  type,
  sheet,
  password = NULL,
  write.res.password = NULL
)
}
\arguments{
\item{data}{object with data to write. For rdata can be a single object or a list of objects.}

\item{file}{file name and path to write to. If missing the object name will be used. No extension is needed, this will be added.}

\item{type}{file type to save data as as a string. If not defined the file name extension will be used or "csv" if no extension given. Options are csv, xlsx, xls, rdata, rdsd, dta.}

\item{sheet}{Optional; for xlsx and xls the sheet name to write data to.}

\item{password}{Password for password-protected workbook.}

\item{write.res.password}{Second password for editing workbook.}
}
\value{
writes data to file
}
\description{
Write data into csv, xlsx, xls, rds, rdata or dta files
}
\examples{

# simplewrite(data, "C:..../file.csv")

# simplewrite(data, "C:..../file.xlsx", sheet = "Data")

# simplewrite(data, "C:..../file.xls", sheet = 4)

}
\author{
Diane Hatziioanou
}
\keyword{save}
\keyword{write,}
