% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlfile.R
\name{sql_nocomments}
\alias{sql_nocomments}
\title{Title Remove comments from SQL files}
\usage{
sql_nocomments(sqlfile = NULL, q = NULL)
}
\arguments{
\item{sqlfile}{sql query file}

\item{q}{query from sql file}
}
\value{
query with comments removed
}
\description{
Title Remove comments from SQL files
}
\examples{
# sqlfile_tables(sqlfile = "sqlfile.sql")
# sqlfile_tables(q = q1)

}
