% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlfile.R
\name{sqldb_retrieve}
\alias{sqldb_retrieve}
\title{Title Import dataset from SQL server into R}
\usage{
sqldb_retrieve(
  name = "dt",
  server,
  database,
  schema_name = NULL,
  dataset,
  columns = FALSE,
  top = FALSE
)
}
\arguments{
\item{name}{name to call retrieved data in environment}

\item{server}{Server holding data}

\item{database}{Database to query}

\item{schema_name}{dataset schema eg "dbo"}

\item{dataset}{table or view dataset name}

\item{columns}{column names}

\item{top}{Optional number of top rows to retrieve}
}
\value{
Dataset contents as defined by arguments from SQL server
}
\description{
Title Import dataset from SQL server into R
}
\examples{
# sqldb_retrieve(name ="dt", server="x", database="W001", dataset ="x_linelist_v1")
# sqldb_retrieve(name ="dt", server="x", database="W001", dataset ="x_linelist_v1",
                    # columns=c("id", "value"), top =100)

}
