% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlfile_query.R
\name{sqlfile_query}
\alias{sqlfile_query}
\title{Title Use sql query file to query database via odbc}
\usage{
sqlfile_query(
  server = load_data$CVD19_Server,
  database = "W126_Covid_OST_Surveillance",
  sqlfile = file.path("R", "SQL", "CVD19_episode_var_QAall.sql"),
  name = "df"
)
}
\arguments{
\item{server}{Server holidng data}

\item{database}{Database to query}

\item{sqlfile}{sql query file}

\item{name}{object to give retrieved data}
}
\value{
retrieves data to .GlobalEnv
}
\description{
Title Use sql query file to query database via odbc
}
\examples{
# sqlfile_query(server = "ser.path", database ="W001", sqlfile  = "query.sql", name = "df")


}
