% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlfile.R
\name{sqlfile_read}
\alias{sqlfile_read}
\title{Title Read sql query file}
\usage{
sqlfile_read(sqlfile)
}
\arguments{
\item{sqlfile}{sql query file}
}
\value{
sql file formatted for use in R
}
\description{
Title Read sql query file
}
\examples{
query <- sqlfile_read("C:/Users/path/query1.sql")

}
