% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlfile.R
\name{sqlfile_tables}
\alias{sqlfile_tables}
\title{Title Extract names of tables from sql query}
\usage{
sqlfile_tables(sqlfile = NULL, q = NULL, nameonly = FALSE)
}
\arguments{
\item{sqlfile}{sql query file}

\item{q}{query from sql file}

\item{nameonly}{Optional; list only table/view name. Default is FALSE}
}
\value{
table names from query
}
\description{
Title Extract names of tables from sql query
}
\examples{
# sqlfile_tables(q = query1)
# sqlfile_tables(sqlfile = "query.sql")

}
