% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PII_manage.R
\name{vectorise_unique}
\alias{vectorise_unique}
\title{Title Format full names from a data.table and return unique names per row.}
\usage{
vectorise_unique(x, namecols)
}
\arguments{
\item{x}{data.table or data.frame with columns containing messy names}

\item{namecols}{column names containing names (forename(s), middlename(s), surname(s)) to process}
}
\value{
list of unique names
}
\description{
Title Format full names from a data.table and return unique names per row.
}
\examples{
 # dt <- data.table(first_name = c("Jane", "Pete", "John Patrick"), surname = c("Brown", "McKenzie", "Patrick McKenzie"))
 # dt$namelist <- vectorise_unique(x = dt, namecols =  c("first_name","surname"))
 # Tidy surname column and return a vector of all names per row
 # dt$names <- lapply(dt$namelist, function(x) paste0(unlist(x), collapse = " "))

}
