% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_any.R
\name{write_any}
\alias{write_any}
\title{Write data to file in various formats with one command}
\usage{
write_any(data, file, type, sheet)
}
\arguments{
\item{data}{object with data to write. For rdata can be a single object or a list of objects.}

\item{file}{file name and path to write to. If missing the object name will be used. No extension is needed, this will be added.}

\item{type}{file type to save data as as a string. If not defined the file name extension will be used or "csv" if no extention given. Options are csv, xlsx, xls, rdata, rdsd, dta.}

\item{sheet}{Optional; for xlsx and xls the sheet name to write data to.}
}
\value{
writes data to file
}
\description{
Write data into csv, xlsx, xls, rds, rdata or dta files
}
\examples{

write_any(data, "C:..../file.csv")

write_any(data, "C:..../file.xlsx", sheet = "Data")

write_any(data, "C:..../file.xls", sheet = 4)

}
\author{
Diane Hatziioanou
}
\keyword{save}
\keyword{write,}
