% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{backup}
\alias{backup}
\title{Title  Backup folder contents where change identified with \code{\link[simpleepi]{simplefilecheck}}}
\usage{
backup(from, to, silent = FALSE)
}
\arguments{
\item{from}{folder path with most up to date data}

\item{to}{folder path where backups stored}

\item{silent}{Optional; disaply messages}
}
\value{

}
\description{
Title  Backup folder contents where change identified with \code{\link[simpleepi]{simplefilecheck}}
}
\examples{
# backup(from ="path1", to = "path2")

}
