% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_vc.R
\name{file_stat_diff}
\alias{file_stat_diff}
\title{Title}
\usage{
file_stat_diff(
  logs = file_stat_log,
  columns,
  logpath = NULL,
  logname = NULL,
  maxup = 1.25,
  maxdown = 0,
  Date = Sys.Date()
)
}
\arguments{
\item{logs}{log created using file_stat_log}

\item{columns}{columns within log with quantified value combinations}

\item{logpath}{Path to log}

\item{logname}{Name of log}

\item{maxup}{Optional; Maximum permissible increase from previous log. Warning recorded if exceeded. Default is 1.25}

\item{maxdown}{Optional; Maximum permissible decrease from previous log. Warning recorded if exceeded. Default is 0.}

\item{Date}{Optional; date to give processed log. Default is system date.}
}
\value{

}
\description{
Title
}
\examples{

}
