% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age_groups.R
\name{ONS_age_groups}
\alias{ONS_age_groups}
\title{Title Calculate ONS populations by age group for any age grouping using ONS populations file from https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationprojections/datasets/tablea21principalprojectionukpopulationinagegroups}
\usage{
ONS_age_groups(populations, age_groups)
}
\arguments{
\item{populations}{object containing ONS populations}

\item{age_groups}{data.frame containing age group categories to calculate populations for}
}
\value{
exports a data.frame with aggregated population numbers by age group
}
\description{
Title Calculate ONS populations by age group for any age grouping using ONS populations file from https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationprojections/datasets/tablea21principalprojectionukpopulationinagegroups
}
\examples{

 # Create age grouping
 # groups <- data.frame(under18 = c(0,17),
           # workingage = c(18,65),
           # over65 = c(66,"90+"),
           # stringsAsFactors = FALSE)
 # ONS_populations_3categories <- ONS_age_groups(ONS, groups)

}
