% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns_manage.R
\name{QC_categories}
\alias{QC_categories}
\title{Title Compare vector contents to a expected categories}
\usage{
QC_categories(data, cat)
}
\arguments{
\item{data}{vector to perform check on}

\item{cat}{categorical values to look for}
}
\value{
returns list containign summary of comparison, whether data fully conforms to categories and whether data has missing of additional categories
}
\description{
Title Compare vector contents to a expected categories
}
\examples{

# check1 <- QC_categories(df$coolumn1, cat = c("low", "medium", "high"))

}
