% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\name{replicates}
\alias{replicates}
\title{Retrieve duplicated rows by one or more columns}
\usage{
replicates(df, columns)
}
\arguments{
\item{df}{data}

\item{columns}{column or vector of columns to look for duplicate combinations in}
}
\value{
Returns duplicated rows within df based on the specified columns
}
\description{
Retrieve duplicated rows by one or more columns
}
\examples{

df_duplicates <- replicates(df, c("Name","DOB"))


}
\author{
Diane Hatziioanou
}
\keyword{duplicates,}
\keyword{replicates}
