% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates_rmv.R
\name{replicates_rmv}
\alias{replicates_rmv}
\title{Remove duplicated rows by one or more columns}
\usage{
replicates_rmv(df, columns)
}
\arguments{
\item{df}{data}

\item{columns}{column or vector of columns to look for duplicate combinations in}
}
\value{
Returns df without duplicated rows  based on the specified columns
}
\description{
Remove duplicated rows by one or more columns
}
\examples{

df_exclude_all_duplicates <- replicates_rmv(df, c("ID1","ID2"))


}
\author{
Diane Hatziioanou
}
\keyword{duplicates,}
\keyword{replicates}
