% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleaggregate.R
\name{simpleaggregate}
\alias{simpleaggregate}
\title{Title}
\usage{
simpleaggregate(dt, idcol, disregard, prefkeep)
}
\arguments{
\item{dt}{data}

\item{idcol}{column with unique identifiers to aggregate by}

\item{disregard}{Optional; values to remove from aggregation eg NA, NULL, 0, -1. Default is NA.}

\item{prefkeep}{Optional; Numeric value representing which comma separated value to keep where aggregation results in more than one comma separated values. Default is to keep all values.}
}
\value{

}
\description{
Title
}
\examples{

}
