% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_manage.R
\name{simpleaggregate}
\alias{simpleaggregate}
\title{Title Combine rows sharing unique identifier to save all unique data to 1 row per identifier}
\usage{
simpleaggregate(
  dt,
  idcol,
  reformat = TRUE,
  disregard = c(NA, ""),
  prefkeep = FALSE,
  col = NULL
)
}
\arguments{
\item{dt}{data}

\item{idcol}{column with unique identifiers to aggregate by}

\item{reformat}{Optional; TRUE (default) reformats to original column classes. FALSE returns all as character. Use FALSE to avoid losing different dates per ID.}

\item{disregard}{Optional; cell values to remove prior to aggregation
eg NA, NULL, 0, -1. Default; NA, "".}

\item{prefkeep}{Optional; Keep 1st record value where duplicate record values differ. Default; FALSE.}

\item{col}{Optional; Columns to process with prefkeep. "all" will process all columns}
}
\value{
Returns data.frame aggregated to one row per idcol.
}
\description{
Title Combine rows sharing unique identifier to save all unique data to 1 row per identifier
}
\examples{
messy <- data.frame(ID = c(1,1,2,3),
data <- c("red", "yellow", "red", "green"))
clean <- simpleaggregate(
   dt = messy,
   idcol = "ID",
   disregard=c(NULL,NA,""))
clean <- simpleaggregate(dt = messy,
   idcol = "ID",
   disregard=c(NULL,NA,""),
   prefkeep = 1,
   col = "all")

}
