% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_manage.R
\name{simpleappendenv}
\alias{simpleappendenv}
\title{Combine tables from environment into one table by appending rows}
\usage{
simpleappendenv(x, name = "DF")
}
\arguments{
\item{x}{data.frame or data.table to append}

\item{name}{name to call database of appended x}
}
\value{
object named with name argument in .GlobalEnv with data from x appended at end
}
\description{
Combine tables from environment into one table by appending rows
}
\examples{
for (v in 1:3){
  df <- data.frame(a = stats::runif(3), b = c("a", "b", "c"))
  df$version <- v
  simpleappendenv(x = df, name = "DF")
}

}
