% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplelog.R
\name{simplelog_end}
\alias{simplelog_end}
\title{Title Write data and end connection to history and log files}
\usage{
simplelog_end(path, history_file, output_file)
}
\arguments{
\item{path}{Path to history and output files}

\item{history_file}{file name for history}

\item{output_file}{file name for output}
}
\value{
Closes connection to history and output files
}
\description{
Title Write data and end connection to history and log files
}
\examples{
# simplelog_end()

}
