% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplelog.R
\name{simplelog_start}
\alias{simplelog_start}
\title{Title Start connection to history and log files}
\usage{
simplelog_start(path, history_file, output_file)
}
\arguments{
\item{path}{Optional path to history and output files. Default is working directory.}

\item{history_file}{Optional file name for history. Default is file.path(path, paste0(date,"_Rhistory.txt"))}

\item{output_file}{Optional file name for output. Default is file.path(path, paste0(date, "_Rlog.txt"))}
}
\value{
Opens connection to history and output files
}
\description{
Title Start connection to history and log files
}
\examples{
# simplelog_start()
# path = "getwd()"
# history_file <- file.path(path, paste0(Sys.Date(),"Rhistory.txt"))
# output_file <- file.path(path, paste0(Sys.Date(),"Rlog.txt"))
# simplelog_start(path, history_file, output_file)

}
