% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_manage.R
\name{simpleopen}
\alias{simpleopen}
\title{Open path using File Explorer or open named file}
\usage{
simpleopen(path)
}
\arguments{
\item{path}{Path to folder or file to open}
}
\value{
Opens a File Explorer window at the given path or file
}
\description{
Open path using File Explorer or open named file
}
\examples{
# dir.create("test")
# simpleopen("test")

}
