% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlfile.R
\name{sqldb_retrieve}
\alias{sqldb_retrieve}
\title{Title Import full dataset from SQL server into R}
\usage{
sqldb_retrieve(name = "dt", server, database, schema_name = NULL, dataset)
}
\arguments{
\item{name}{name to call retrieved data in environment}

\item{server}{Server holding data}

\item{database}{Database to query}

\item{schema_name}{dataset schema eg "dbo"}

\item{dataset}{table or view dataset name}
}
\value{
Full defined databset from SQL server
}
\description{
Title Import full dataset from SQL server into R
}
\examples{
# sqldb_retrieve(name ="dt", server="x", database="W001", dataset ="x_linelist_v1")

}
