% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearweek.R
\name{yearweek}
\alias{yearweek}
\title{Title Format a date to isoyear-isoweek}
\usage{
yearweek(x)
}
\arguments{
\item{x}{Date}
}
\value{
Returns x in format YYYYWW where WW is the isoweek and YYYY is the isoyear.
}
\description{
Title Format a date to isoyear-isoweek
}
\examples{
yw <- simpleyearweek(date)
x$yw <- simpleyearweek(x$date)

}
